LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.quadraticCurveTo.nonfinite.worker-expected.txt


PASS quadraticCurveTo() with Infinity/NaN is ignored
